Move i18n, account list and version list to their own files
